+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - added store type setting to myems-admin
 - added optional redis database to myems-api
+- added phone to tbl_users in database
+- added login by phone to myems-api and myems-web
+
 ### Changed
 - udpated space settings in myems-admin
 ### Fixed